refactor: Update module imports and paths for consistency and clarity Bug: with the modules detections cant define the database modules
