+++ src/database/index.ts
@@ -1,0 +1,3 @@
+export * from './database.module';
+export * from './database.service';
+export * from './migrate';
